refactor : Voucher class -> interface로 리팩터링
