+++ src/main/java/org/prgrms/kdt/customer/controller/CreateCustomerRequest.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.customer;
+package org.prgrms.kdt.customer.controller;
 
 public record CreateCustomerRequest(String email, String name) {
 }
