+++ src/main/java/org/prgrms/kdt/customer/controller/CustomerController.java
@@ -1,5 +1,7 @@
-package org.prgrms.kdt.customer;
+package org.prgrms.kdt.customer.controller;
 
+import org.prgrms.kdt.customer.service.CustomerService;
+import org.prgrms.kdt.customer.model.Customer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
