+++ src/main/java/org/prgrms/kdt/customer/service/CustomerServiceImpl.java
@@ -1,6 +1,7 @@
-package org.prgrms.kdt.customer;
+package org.prgrms.kdt.customer.service;
 
-import org.prgrms.kdt.aop.TrackTime;
+import org.prgrms.kdt.customer.repository.CustomerRepository;
+import org.prgrms.kdt.customer.model.Customer;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
